+++ src/main/java/com/google/error_analyzer/servlets/DocCountServlet.java
@@ -8,8 +8,10 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
-package com.google.error_analyzer.backend;
+package com.google.error_analyzer.servlets;
 
+import  com.google.error_analyzer.backend.LogDao;
+import  com.google.error_analyzer.backend.LogDaoHelper;
 import  com.google.error_analyzer.data.constant.FileConstants;
 import com.google.gson.Gson;
 import java.io.IOException;
