+++ src/test/java/com/google/error_analyzer/backend/DocCountServletTest.java
@@ -10,7 +10,7 @@ See the License for the specific language governing permissions and
 limitations under the License.*/
 package com.google.error_analyzer;
 
-import com.google.error_analyzer.backend.DocCountServlet;
+import com.google.error_analyzer.servlets.DocCountServlet;
 import com.google.error_analyzer.backend.LogDao;
 import java.io.*;
 import java.util.*;
